PRSD-1109: Updates CheckOccupancyAnswersPage to match design
