+++ src/main/resources/messages.properties
@@ -784,6 +784,7 @@ forms.update.numberOfPeople.fieldSetHeading=Update how many people live in your 
 forms.update.licensingType.fieldSetHeading=Update the type of licensing you have for your property
 forms.update.checkLicensing.update.summaryName=You have updated the property licence
 forms.update.checkLicensing.remove.summaryName=You have removed this property\u2019s licence
+forms.update.checkOccupancy.summaryName=You have updated whether the property is occupied by tenants
 forms.update.warning=You must not provide any false or misleading information on the database. By submitting this information, you agree it\u2019s accurate to the best of your knowledge at the time of updating.
 
 forms.areYouSure.propertyDeregistration.fieldSetHeading=Are you sure you want to delete {0,,address} from the database?
